+++ tests/controllers/__init__.py
@@ -10,7 +10,7 @@ class BaseTestCase(TestCase):
 
     def create_app(self):
         logging.getLogger('connexion.operation').setLevel('ERROR')
-        app = connexion.App(__name__, specification_dir='../det/swagger/')
+        app = connexion.App(__name__, specification_dir='../../det/swagger/')
         app.app.json_encoder = JSONEncoder
         app.add_api('swagger.yaml')
         return app.app
