+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
+envlist = py36
 envlist = py35
-
 [testenv]
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
@@ -8,7 +8,7 @@ setenv=
      atlas_login = atlas_login
      atlas_password = atlas_password
 install_command = 
-     pip install --upgrade --allow-all-external {packages}
+     pip install --upgrade {packages}
 commands=
-     pytest \
-      []
+     pytest tests/controllers
+     pytest --ignore=tests/controllers
